+++ packages/tools/geo/rollup.config.js
@@ -22,6 +22,9 @@ const analyzer = analyze({
 const banner = makeBanner(pkg);
 const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
 const input = pkg.module;
+const removeComments = cleanup({
+	extensions: ['js', 'mjs']
+});
 const treeshake = {
 	annotations: true,
 	moduleSideEffects: [
@@ -44,7 +47,7 @@ const cjsConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		cleanup(),
+		removeComments,
 	],
 	treeshake
 };
@@ -62,7 +65,7 @@ const browserConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		cleanup(),
+		removeComments,
 		buble({objectAssign: 'Object.assign'}),
 	],
 	treeshake