+++ packages/tools/request/rollup.config.js
@@ -22,6 +22,9 @@ const analyzer = analyze({
 const banner = makeBanner(pkg);
 const external = pkg.peerDependencies && Object.keys(pkg.peerDependencies) || [];
 const input = pkg.module;
+const removeComments = cleanup({
+	extensions: ['js', 'mjs']
+});
 const treeshake = {
 	annotations: true,
 	moduleSideEffects: [
@@ -41,7 +44,7 @@ const cjsConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		cleanup(),
+		removeComments,
 	],
 	treeshake
 };
@@ -59,7 +62,7 @@ const browserConfig = {
 	plugins: [
 		resolve(),
 		commonjs(),
-		cleanup(),
+		removeComments,
 		buble(),
 	],
 	treeshake